Make `npm install -g jest-cli` work reasonably well
Previously, if you were to `npm install -g jest-cli`, life would suck if
you had an old version of jest installed globally...but your tests only
work properly with new versions.

This makes the jest binary try to be smart and look for a jest binary
that might already exist in CWD project -- and if one is found, use that
to run the tests instead of the global version.