Make useStderr a proper config value
I've used `config.useStderr` internally to divert stdout to stderr
so we can talk to the caller via json. Instead of having a one-off
config option, I'm making this an actual option with a consistent
behaviour.

Had to convert the `InstanbulTestReporter` into a proper class
so it can do proper inheritence.