Add module.parent to modules loaded by haste
Summary: `module.parent` is basically `arguments.callee` for node. I'm adding it to jest so that modules don't think they are run standalone as a script. This fixes jest 0.6.0 on react-native because of the `rc` module: https://github.com/dominictarr/rc/blob/master/index.js#L56

I will add more stuff to all of this soon but with the impending doom of the HasteModuleLoader and the rewrite, there is no point in doing more than the minimum necessary to support jest's clients :)

public

Reviewed By: voideanvalue

Differential Revision: D2573421

fb-gh-sync-id: ef049b519d947112b21458056c2498549a961c7d