Remove deprecated built-in modules and start cleaning up HasteModuleLoader
Summary: Removing these APIs internally was actually quicker than I expected (yay codemods). Now I can clean-up some things inside of the module loader and remove this deprecated functionality. Yay! I got a little bit carried away when I touched this file and I should have probably split this up into multiple diffs, but it is a hackathon and I hope you'll forgive me. This is the first step of refactoring the module loader so we can start replacing node-haste.

This diff

* Removes ~250 files of code from the module loader and is focused on making the file a little bit more readable
* Removes the concept of virtual `builtinModules` and removes access of `mock-modules`, `mocks` and `node-haste`. This shouldn't be a thing. It adds `createRuntimeFor(path)` which is similar to how the jest runtime was accessed before. Previously this function was re-created every time `resetModuleRegistry` was called and a new runtime was (potentially) being created. This is pointless because the global reference to `jest` inside of a test stayed the same anyway. `require('jest-runtime')` is not possible any more, `jest` is now the only way to access the jest API, as it should be. All of these used to be Facebook internal private APIs.
* Converted the code into using classes, let and const
* Uses `Object.keys` and proper `hasOwnProperty` access
* Removes comments or "type" annotations that are not useful and make it harder to skim over the file.
* Reformats code to fit into the current code style of jest; fixed all lint issues.

Note that this entire file is going to be refactored into different pieces soon, and this is just the first of many steps. Functionally it is identical and all tests are passing.

public

Reviewed By: voideanvalue

Differential Revision: D2600019

fb-gh-sync-id: 009e8580e75f319e87e85b5e8ea919b9cc80df90