Fix test setup for module loader
Summary: The jest tests have always been disabled on internal test runs on sandcastle because they tried to write to `.haste_cache` inside `scripts/jest`. This is a long standing issue that came up again because I moved jest to a different location which made the test infra think they are new (and I got assigned a bunch of tasks). I knew about this before and attempted to fix it but couldn't figure out what was wrong. It turns out that testing jest with jest makes everything a little bit more confusing:

* HasteModuleLoader tests write and read directly from disk. They create their own module loader (in a jest test that uses a different instance of the loader) and the default cache dir is set to `.haste_cache` within the package. This diff changes it to read from a global `CACHE_DIRECTORY` that we can use to point to the appropriate cache dir (internal vs. external). Long term we might actually want to write the test-caches into a separate folder or a virtual file system but I'll leave that up for later when we have replaced the resolver.
* The call to `path.normalize` in utils was incorrect. The `./` needed to be moved inside the call so that a path looking like `/../` can be correctly normalized to `../`. Without this, the change in package.json has to look like `<rootDir>../.haste_cache`, which doesn't make any sense.
* Multiple tests didn't provide a name for the module loader which means that they all tried to read/write to the same cache file (`cache-undefined`, meh). This caused some log-spew and potentially also caused tests to not work properly in parallel

public

Reviewed By: voideanvalue

Differential Revision: D2602911

fb-gh-sync-id: fcc6c501c54c196d5f9242cb5e11f894bb8d3c8e