Use chalk instead of managing our own colorization.
Summary:
This is arguably slightly nicer but also avoids having to lookup config.noHighlight, instead just turning that off in chalk at the top level when we see that flag.

The other thing this provides is that we never have to do resets ourselves, which resulted in a bug recently (which was why I looked into this at all).
Closes https://github.com/facebook/jest/pull/639

Reviewed By: svcscm

Differential Revision: D2746396

fb-gh-sync-id: c66a193d55d0d04c9550d16a4970c5b1fabfcd4a