Clarify the difference between jest.runAllTimers and jest.runAllTicks
Summary:
I read the docs for ```runAllTicks``` and ```runAllTimers``` ~3 times before I finally realized one was for the *micro* task queue and the other is for the *macro* task queue. I'm sure other people have done the same thing, so I think it would help to bold the "micro" and "macro" text.
Closes https://github.com/facebook/jest/pull/653

Reviewed By: svcscm

Differential Revision: D2765491

fb-gh-sync-id: 079b31e86f6fc72faa31e7cb6c90d4fb88637efc