Bind all process event functions
Summary:Not all process event functions are correctly bind to the original process object. Currently only process.on is bind correctly. This pull requests binds all EventEmitter functions to the correct object.

I am not 100% sure, but maybe this fix could be applied to all process functions instead of only the EventEmitter ones? This could easily be done as following:

```js
for (var func in process)
{
  if (typeof process[func] == 'function')
  {
    global.process[func] = process[func].bind(process);
  }
}
```
Closes https://github.com/facebook/jest/pull/738

Differential Revision: D2970587

fb-gh-sync-id: 41f5aacd2775327d5b87900115e1a9e3e5294293
shipit-source-id: 41f5aacd2775327d5b87900115e1a9e3e5294293