Initial implementation of a blog support
Summary:This adds blog post rendering, left nav with the list of blog posts, highlighting the currently selected blog post and highlighting the header links.

I cp all the React blog posts in order to test it out, working great.

Missing pieces:
 - I haven't implemented a listing of all blog posts, the header link opens the most recent one.
 - I haven't implemented truncation of the links on the left and the all page

Those are needed when we'll have > 1 blog post, which is not now :)

bypass-lint

![screen shot 2016-03-02 at 3 55 09 pm](https://cloud.githubusercontent.com/assets/197597/13480193/011abff8-e091-11e5-8421-677217f383ff.png)
Closes https://github.com/facebook/jest/pull/760

Differential Revision: D3006427

fb-gh-sync-id: 0e635247732f6187f9255ea92d59701f1a467af4
shipit-source-id: 0e635247732f6187f9255ea92d59701f1a467af4