Ignore all node_modules directories in existance
Summary:The problem was that the ignore patterns we had weren't right when passed to `glob`. This meant that `glob` was returning thousands of files it shouldn't have which eslint was then checking again and finally ignoring there.

We don't ever want to lint anything in any `node_modules` directory so just ignore all of them all the time.

After a simple `npm install` inside `examples/async` this brought be down from 17s back to what I had before that install, ~5s.
Closes https://github.com/facebook/jest/pull/780

Differential Revision: D3026152

fb-gh-sync-id: bf92a7d28b50d01e2a8b6dbc654e488fa4163ebe
shipit-source-id: bf92a7d28b50d01e2a8b6dbc654e488fa4163ebe