Change preprocessorIgnorePathPatterns for better react-native support
Summary:Discussion: https://www.facebook.com/groups/reactnativeoss/permalink/1528961550733807/

I would like `babel-jest` and `jest`'s setup process to be simpler. It turns out that react-native plugin developers commonly ship plugins without precompiling them with babel, leaving it to react native's packager.

This changes the `preprocessorIgnorePathPatterns` config option to default to `['/node_modules/']` or an empty array when react-native is detected as part of the project. Further, I removed the `node_modules` check in babel-jest, because the preprocessor is now not called when the file is a node_module. This has the nice side-effect of *not* copying every single node_module JS file into the cache and the files will now be read from the cache directly! For users of babel-jest, nothing will change. For people using react-native, they'll now be able to use `babel-jest` directly. For people who do not use a preprocessor, this change doesn't have an effect either. It does however change Jest for people who use
Closes https://github.com/facebook/jest/pull/779

Reviewed By: yungsters

Differential Revision: D3025485

fb-gh-sync-id: 56254634383d28cd9952adcfb0081ab9ec4f0fa6
shipit-source-id: 56254634383d28cd9952adcfb0081ab9ec4f0fa6