Support Jasmine2 pending tests
Summary:As of Jasmine2 there is the possibility to determine whether a task is pending or not. This allows for those pending tasks to output with a custom yellow icon (\u25CB small white circle &#x25CB;) in verbose mode and yellow pending information in the summary line.

You can set tests as pending in Jasmine2 by using `xit()`, `xdescribe()`, `pending()`, or by not passing a function to `describe()`, `it()`, `pit()`, `fit()`, etc.

```
describe('AwesomeClass', () => {
  xit('going to write tests one day');
  pit('going to write tests one day');
  it('going to write tests one day');
  fit('going to write tests one day');
});
```

This can not work for Jasmine1 because it does not output the tasks with any form of pending information.

fixes #784
Closes https://github.com/facebook/jest/pull/788

Differential Revision: D3039885

fb-gh-sync-id: 3e502bb0af47c7e1854f4b7d2a277c7ec29ad102
shipit-source-id: 3e502bb0af47c7e1854f4b7d2a277c7ec29ad102