Proper haste package support
Summary:I didn't make a distinction between regular modules and haste packages. This means that when looking at a node_module that is also using `providesModule` (like react-native), it will set `react-native` to a haste-package based on its `package.json`. Previously, react-native was just added as a real module so `require('react-native')` resolves to the `package.json` file instead of the `main` file it should resolve to normally.

This diff fixes it and adds some tests for this behavior.

The way this fix works is by splitting up modules and haste packages into separate maps. The resolution algorithm in Jest works like this:

* modules
* node modules resolution
* haste packages

Without this diff, `require('react-native')` would resolve in the `modules` stage and return the path to the `package.json`. With this diff, it will fall through the first step and use the regular node modules resolution.

public

Reviewed By: voideanvalue

Differential Revision: D3044486

fb-gh-sync-id: 7bdd71e434ab3ef2d9854da109900ec0eb660594
shipit-source-id: 7bdd71e434ab3ef2d9854da109900ec0eb660594