Fix unmock cache initialization.
Summary:
We removed global mock caches in #934 but for some reason I forgot to move this cache in the constructor. We only unmock the setup files a single time which isn't useful for subsequent tests. This makes it so we do this every time. It shouldn't have a big perf impact.
Closes https://github.com/facebook/jest/pull/942

Differential Revision: D3229346

fb-gh-sync-id: f58a2b5cc65160c2d0a63e874bb9257d3218f1a6
fbshipit-source-id: f58a2b5cc65160c2d0a63e874bb9257d3218f1a6