Wrap git/hg detectors in try-catch. See #905.
Summary:
In #905 it was reported that people without mercurial had some trouble with `jest -o`. I fixed that initially but on some system it seems like `childProcess.spawn` can throw. It doesn't seem to happen on OS X locally but based on what I read about `childProcess.spawn` I do believe it can throw. I think it is best to wrap it in a try-catch now – the worst thing that can happen is to report back as not a git or not an hg repo, which it won't be if the commands fail.

I also took a note to write an integration test for this, now that we have integration tests. I will add a few in a follow up.
Closes https://github.com/facebook/jest/pull/941

Reviewed By: kentaromiura

Differential Revision: D3229317

fb-gh-sync-id: 34bffb7f256924b9766f6b12ba9c81ac663d8eba
fbshipit-source-id: 34bffb7f256924b9766f6b12ba9c81ac663d8eba