Add `jest-resolve` to resolve modules based on the Facebook module resolution algorithm.
Summary:
This refactors Runtime.js and TestRunner.js and pulls out the resolution code into a new module `jest-resolve` which now encodes Facebook's module resolution algorithm. It provides the second half of node-haste and together with `jest-haste-map` can be seen as a full replacement.

There is very little new code in this diff. Most code is existing code that was lifted from another module and brought together. The next step after this diff will be to see if it makes sense to fork the `resolve` module and speed up how it works.

It doesn't come with its own tests (yet) because its functionality is completely covered by all the tests in Runtime.js. I have also cleaned up all of the Runtime tests and added a new `createRuntime` helper for testing, which should make it much easier from now on to change how an instance of Runtime is constructed for tests.

Closes https://github.com/facebook/jest/pull/1007

Reviewed By: dmitriiabramov

Differential Revision: D3302805

fbshipit-source-id: d1c96de7ed09e36aa1b2b5a4656260e79d1fb3e2