Remember to escape the backslash in RegEx
Summary:
Jest can't find tests on Windows because of a bad RegEx.

**Steps to reproduce:**
npm install
cd examples
cd getting_started
npm install
npm test

**Results:**

```
> @ test C:\dev\jest\examples\getting_started
> jest

Using Jest CLI v12.1.0, jasmine2
No tests found for "".
```

**Expected results:**
Tests should be found

**Solution:**
The **pathToRegex** function in **TestRunner.js** replaces forward slashes with the directory delimiter, which is a backslash in Windows.  However the backslash happens to be an escape character in RegEx, so the RegEx produced by this substitution is bogus and paths aren't evaluated properly.  The solution is to use a double backslash.  A similar function already exists in the **jest-util** package, but it gets this right.  All I'm doing in this PR is copying the jest-util function to TestRunner.
Closes https://github.com/facebook/jest/pull/1044

Differential Revision: D3339071

fbshipit-source-id: 54fcec8bafd67fde77968debc39952090c2be268