Make options parsing more strict
Summary:
optimist is old, yargs is new.

optimist truncates the help output in the event that `check()` throws an error, meaning the user probably doesn't have any idea what went wrong or why.

yargs shows the untruncated help output.

This is an RFC because we run Jest in a bunch of places and not all will have access to yargs.
Closes https://github.com/facebook/jest/pull/1043

Reviewed By: tomocchino

Differential Revision: D3345323

fbshipit-source-id: de3a453c499a53529227408b2b7a3183d2a1f11f